+++ db/migrate/20220804093319_add_ukprn_to_project.rb
@@ -1,0 +1,5 @@
+class AddUkprnToProject < ActiveRecord::Migration[7.0]
+  def change
+    add_column :projects, :trust_ukprn, :integer, null: false
+  end
+end