+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_07_26_143808) do
+ActiveRecord::Schema[7.0].define(version: 2022_08_04_103138) do
   create_table "actions", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.string "title", null: false
     t.integer "order", null: false
@@ -21,12 +21,24 @@ ActiveRecord::Schema[7.0].define(version: 2022_07_26_143808) do
     t.index ["task_id"], name: "index_actions_on_task_id"
   end
 
+  create_table "notes", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
+    t.text "body"
+    t.uuid "project_id"
+    t.uuid "user_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["project_id"], name: "index_notes_on_project_id"
+    t.index ["user_id"], name: "index_notes_on_user_id"
+  end
+
   create_table "projects", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.integer "urn", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.uuid "delivery_officer_id"
     t.uuid "team_leader_id", null: false
+    t.integer "trust_ukprn", null: false
+    t.date "target_completion_date", null: false
     t.index ["delivery_officer_id"], name: "index_projects_on_delivery_officer_id"
     t.index ["team_leader_id"], name: "index_projects_on_team_leader_id"
     t.index ["urn"], name: "index_projects_on_urn"
@@ -60,6 +72,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_07_26_143808) do
   end
 
   add_foreign_key "actions", "tasks"
+  add_foreign_key "notes", "projects"
+  add_foreign_key "notes", "users"
   add_foreign_key "projects", "users", column: "delivery_officer_id"
   add_foreign_key "projects", "users", column: "team_leader_id"
   add_foreign_key "sections", "projects"