+++ spec/support/academies_api_helpers.rb
@@ -1,7 +1,7 @@
 module AcademiesApiHelpers
-  def mock_successful_api_responses(urn:)
+  def mock_successful_api_responses(urn:, ukprn:)
     mock_successful_api_establishment_response(urn:)
-    mock_successful_api_conversion_project_response(urn:)
+    mock_successful_api_trust_response(ukprn:)
   end
 
   def mock_successful_api_establishment_response(urn:, establishment: nil)
@@ -14,13 +14,13 @@ module AcademiesApiHelpers
     allow(AcademiesApi::Client).to receive(:new).and_return(test_client)
   end
 
-  def mock_successful_api_conversion_project_response(urn:, conversion_project: nil)
-    conversion_project = build(:academies_api_conversion_project) if conversion_project.nil?
+  def mock_successful_api_trust_response(ukprn:, trust: nil)
+    trust = build(:academies_api_trust) if trust.nil?
 
-    fake_result = AcademiesApi::Client::Result.new(conversion_project, nil)
+    fake_result = AcademiesApi::Client::Result.new(trust, nil)
     test_client = AcademiesApi::Client.new
 
-    allow(test_client).to receive(:get_conversion_project).with(urn).and_return(fake_result)
+    allow(test_client).to receive(:get_trust).with(ukprn).and_return(fake_result)
     allow(AcademiesApi::Client).to receive(:new).and_return(test_client)
   end
 end