+++ CHANGELOG.md
@@ -11,6 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The project information view displays the school type, phase of education,
   and age range.
 - The project information view displays the team lead.
+- Users must provide the UKPRN of the incoming trust when creating a project.
+- The project information view has a Trust details section, which displays the
+  incoming trust name, UKPRN, and Companies House number.
+- Users must provide the target completion date when creating a project.
+- Users can add notes to a project, and view previous project notes.
 
 ## [Release 1][release-1]